backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
defusedxml==0.7.1
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
httpie==2.5.0
idna==3.2
Pillow==8.3.2
psycopg2-binary==2.9.1
Pygments==2.10.0
PyJWT==2.1.0
PySocks==1.7.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.6
webcolors==1.11.1
